requirements.txt
mcp==1.0.0
pydantic==2.5.0
langchain==0.1.0
langchain-community==0.0.10
python-dotenv==1.0.0
geopy==2.4.0
httpx==0.25.0
pytest-asyncio==0.21.0
